+++ internal/server/hub.go
@@ -2180,6 +2180,8 @@ func (h *hub) run() {
 				h.onPlaysound(players, e)
 			case evStopsound:
 				h.onStopsound(players, e)
+			case evTellraw:
+				h.onTellraw(players, e)
 			case evParticleCmd:
 				h.spawnParticles(players, e.dim, e.pid, e.x, e.y, e.z, e.spread, e.speed, e.count)
 			case evBoneMeal: